Create automated_with_Timeseries
This is an automated code to extract snow depth from snow poles. It allows the user to calibrate the code to the user's unique snow pole color (red, orange and yellow work best). The angle of the sun can affect the shade of color that the computer perceives, and there can be day-to-day scatter. The minimum value of each day is the closest to the actual value in preliminary analysis.